refactor: Add ShopCatalog component and update Shop page
Add the ShopCatalog component to display a catalog of products in the Shop page. This improves the functionality and user experience of the Shop page by allowing users to browse and view products. Also, update the Shop page to import and render the ShopCatalog component. This change enhances the overall functionality and organization of the Shop page.